chore: add missing exports. Move SSI-types to types folder, which is more in line with typical ts projects
