+++ lib/evaluation/handlers/abstractEvaluationHandler.ts
@@ -2,7 +2,7 @@ import { InputDescriptor, PresentationDefinition, PresentationSubmission } from 
 import jp from 'jsonpath';
 
 import { Status } from '../../ConstraintUtils';
-import { VerifiableCredential } from '../../verifiablePresentation';
+import { VerifiableCredential } from '../../types';
 import { EvaluationClient } from '../evaluationClient';
 import { HandlerCheckResult } from '../handlerCheckResult';
