+++ lib/evaluation/handlers/inputDescriptorFilterEvaluationHandler.ts
@@ -3,8 +3,8 @@ import Ajv from 'ajv';
 import jp, { PathComponent } from 'jsonpath';
 
 import { Status } from '../../ConstraintUtils';
+import { VerifiableCredential } from '../../types';
 import { JsonPathUtils } from '../../utils/jsonPathUtils';
-import { VerifiableCredential } from '../../verifiablePresentation';
 import { EvaluationClient } from '../evaluationClient';
 import { HandlerCheckResult } from '../handlerCheckResult';
