+++ lib/evaluation/handlers/markForSubmissionEvaluationHandler.ts
@@ -2,7 +2,7 @@ import { PresentationDefinition } from '@sphereon/pe-models';
 import jp from 'jsonpath';
 
 import { Status } from '../../ConstraintUtils';
-import { VerifiableCredential } from '../../verifiablePresentation';
+import { VerifiableCredential } from '../../types';
 import { EvaluationClient } from '../evaluationClient';
 import { HandlerCheckResult } from '../handlerCheckResult';
