+++ lib/evaluation/handlers/subjectIsIssuerEvaluationHandler.ts
@@ -2,8 +2,8 @@ import { Constraints, Optionality, PresentationDefinition } from '@sphereon/pe-m
 import { PathComponent } from 'jsonpath';
 
 import { Status } from '../../ConstraintUtils';
+import { VerifiableCredential } from '../../types';
 import { JsonPathUtils } from '../../utils/jsonPathUtils';
-import { VerifiableCredential } from '../../verifiablePresentation';
 import { EvaluationClient } from '../evaluationClient';
 import { HandlerCheckResult } from '../handlerCheckResult';
