+++ lib/types/SSI.types.ts
