+++ lib/types/index.ts
@@ -3,8 +3,11 @@ import {
   CredentialStatus,
   CredentialSubject,
   InputFieldType,
+  Issuer,
   Presentation,
   Proof,
+  ProofPurpose,
+  ProofType,
   VerifiableCredential,
   VerifiablePresentation,
 } from './SSI.types';
@@ -18,4 +21,7 @@ export {
   Credential,
   Proof,
   InputFieldType,
+  Issuer,
+  ProofType,
+  ProofPurpose,
 };
