+++ lib/utils/jsonPathUtils.ts
@@ -1,11 +1,11 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import jp from 'jsonpath';
 
-import { InputFieldType } from '../verifiablePresentation/SSI.types';
+import { InputFieldType } from '../types/SSI.types';
 
 export class JsonPathUtils {
   /**
-   * @param verifiableCredential: a vc object can be found in verifiablePresentation.verifiableCredential[i]
+   * @param verifiableCredential: a vc object can be found in types.verifiableCredential[i]
    * @param paths: paths that can be found in Field object
    * @return a result object containing value of the correct path in the verifiableCredential and the correct path
    * @example(success result): if you call this method with 1. verifiableCredential:
