+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@sphereon/pe-js",
-  "version": "0.5.0-unstable.1",
+  "version": "0.5.0-unstable.2",
   "description": "A Typescript implementation of the DIF Presentation Exchange specification.",
   "main": "dist/main/index.js",
   "module": "dist/module/index.js",
